+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 - **feat**: interactive problem support
 - **feat**: add max inline data length configuration
 - **feat**: implement bulk import from CPH
+- **feat**: add short name settings for competitive programming platforms
 
 ## 0.1.9